+++ CHANGES.md
@@ -1,8 +1,8 @@
 **Added:**
-* added more stringtable test cases
+* a text area for syntax highlighting
 
 **Changed:**
-* none
+* made open projects in project init window be selected by default when there are projects in workspace
 
 **Fixed:**
 * none